Classes extraction & composer autoload
